BREAKING CHANGE - removes register_driver_adapter
adds row_class declaration to build dataclasses or named tuple instances
for select queries.
introduce QueryParser and QueryDatum to separate the parsing of sql dir
and files, from the building of methods on the Queries object. This
should allow for easier addition of features going forward. The driver
adapter is now available by being injected into the Queries class,
rather than by closure scope when the methods are made.