Introduce a QueryFn protocol
This is to avoid the '# type: ignore' on the sql attribute set to
queries' functions. We then declare _make_sync_fn() and _make_async_fn()
to return this type value. In order to avoid a mypy error when setting
the 'sql' attribute to the (plain) 'fn' function, we cast this value to
a QueryFn.

We also add a _query_fn() helper function for the wrapping logic.