Use contextlib.asynccontextmanager instead home-made aioctxlib
This adds a restriction on Python version >= 3.7 for async drivers. This
is checked in aiosql.py when building the _ADAPTERS mapping by only
inserting asyncpg and aiosqlite for Python version >= 3.7. (We have to
'# type: ignore' the _ADAPTERS.update() statement because mypy complains
for some reason.)

In tests of async drivers we add a 'pytestmark' that will skip tests in
those modules for Python version < 3.7. In test_loading.py, we now use
the 'sqlite3' driver so that this can be run on Python 3.6 (these tests
do not depend on an async driver apparently).

In travis config, we now only run mypy on Python >= 3.7, otherwise we'll
get import errors when checking aiosqlite/asyncpg adapters modules.

We finally update docs' index to document the Python version
requirement.