Apply B905 `flake8-bugbear`/`ruff` recommendations
B905: zip() without an explicit strict= parameter set. strict=True causes the resulting iterator to raise a ValueError if the arguments are exhausted at differing lengths. The strict= argument was added in Python 3.10, so don’t enable this flag for code that should work on <3.10. For more information: https://peps.python.org/pep-0618/

Co-authored-by: Jacob Coffee <jacob@z7x.org>