Add aiosql.adapters module to built packages
When the package is built from git, it works because setuptools-scm
registers a setuptools.file_finders entry point, which provides
setuptools with a list of all files managed by git, and setuptools
adds them to built packages (sdist and wheel).

However, when building from a tarball (generated by git archive) it
does not work, since there are no files managed by VCS. Adding the
package explicitly helps for this case.